add assertions
Add a few assertions to launcher and parser.

Combine Launcher and Main class to streamline application flow.
