+++ src/main/java/chat/utils/Parser.java
@@ -269,6 +269,8 @@ public class Parser {
      * @throws ChatException to pass error message along to CLI.
      */
     public static Command parse(String input) throws ChatException {
+        assert input != null : "Input should not be blank";
+        
         String[] parsedInput = parseInput(input);
         String command = parsedInput[0].trim();
         String args = parsedInput[1].trim();
